requirements.txt
pandas==1.5.3
pyyaml==6.0
pytest==7.4.0
